requirements/base.txt
requests==2.31.0
PyYAML==6.0.1
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4
tqdm==4.66.1
loguru==0.7.2
pydantic==2.6.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.1
docker==6.1.3
cryptography==41.0.7
pytest==7.4.4
